TAPP: Updated our main/data/params to utm
Created a new search area file and varified that gdal_utils was
correctly trimming it by manually selecting a feature, trimming,
than comparing the two raster layers in qgis. Update the poses
file to use utm and gave it the corners of the trimmed tif. Also
snuck in some minor styling changes to tapp.py